+++ src/components/Auth/AuthForm.tsx
@@ -9,11 +9,11 @@ import Container from "@material-ui/core/Container";
 import { useHistory } from "react-router-dom";
 import { Field, Form, Formik } from "formik";
 import { useDispatch, useSelector } from "react-redux";
-import { authActions } from "../../features/Login/authSlice";
+import { authActions } from "../../app/reducers/authSlice";
 import { LoginHandler } from "../../app/api/authApi";
 import type { RootState } from "../../app/store";
 import * as yup from "yup";
-import { FormikInput } from "../custom-form/FormikInput";
+import { FormikInput } from "../../common/Form/FormikInput";
 import { useEffect, useState } from "react";
 // main
 interface initialValuesType {
