+++ src/pages/Settings.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Settings = () => {
+  return <div></div>;
+};
+
+export default Settings;
